Data Applications
Some applications on a vaccination data of the countries

The applications that wanted was:

Data Question 1: Write a code that finds the blanks in the daily vaccination column and pastes the minimum vaccination number of the country to the blank. If there is no daily vaccination data of the country, than fill the blank with "0".

Data Question 2: Write a code that finds the median of the daily vaccination numbers of each country. Than print the top 3 highest median of these numbers.

Data Question 3: Write a code that finds the sum of the daily vaccinations just commited in '1/6/2021'

You can check the codes in codespaces